requirements.txt
Django
django-extensions
djangorestframework
psycopg2
envs
pytest==3.2.1
requests
beautifulsoup4
unittest-tap-reporting==0.2
unittest-xml-reporting==2.1.0
urllib3==1.21.1
future==0.16.0
jsondiff==1.1.1
nose==1.3.7
pytest==3.2.1
python-dateutil
